+++ driver/web/docs/gen/gen_types.go
@@ -232,15 +232,16 @@ type APIKey struct {
 
 // Account defines model for Account.
 type Account struct {
-	AppOrg      *ApplicationOrganization `json:"app_org,omitempty"`
-	AuthTypes   *[]AccountAuthType       `json:"auth_types,omitempty"`
-	Devices     *[]Device                `json:"devices,omitempty"`
-	Fields      *AccountFields           `json:"fields,omitempty"`
-	Groups      *[]AppOrgGroup           `json:"groups,omitempty"`
-	Permissions *[]Permission            `json:"permissions,omitempty"`
-	Preferences *map[string]interface{}  `json:"preferences,omitempty"`
-	Profile     *Profile                 `json:"profile,omitempty"`
-	Roles       *[]AppOrgRole            `json:"roles,omitempty"`
+	AppOrg        *ApplicationOrganization `json:"app_org,omitempty"`
+	AuthTypes     *[]AccountAuthType       `json:"auth_types,omitempty"`
+	Devices       *[]Device                `json:"devices,omitempty"`
+	Fields        *AccountFields           `json:"fields,omitempty"`
+	Groups        *[]AppOrgGroup           `json:"groups,omitempty"`
+	Permissions   *[]Permission            `json:"permissions,omitempty"`
+	Preferences   *map[string]interface{}  `json:"preferences"`
+	Profile       *Profile                 `json:"profile,omitempty"`
+	Roles         *[]AppOrgRole            `json:"roles,omitempty"`
+	SystemConfigs *map[string]interface{}  `json:"system_configs"`
 }
 
 // AccountAuthType defines model for AccountAuthType.
@@ -287,8 +288,8 @@ type AppOrgGroup struct {
 
 // AppOrgPair defines model for AppOrgPair.
 type AppOrgPair struct {
-	AppId *string `json:"app_id"`
-	OrgId *string `json:"org_id"`
+	AppId string `json:"app_id"`
+	OrgId string `json:"org_id"`
 }
 
 // AppOrgRole defines model for AppOrgRole.
@@ -509,6 +510,7 @@ type PartialAccount struct {
 	Params         *map[string]interface{} `json:"params"`
 	Permissions    []Permission            `json:"permissions"`
 	Roles          []AppOrgRole            `json:"roles"`
+	SystemConfigs  *map[string]interface{} `json:"system_configs"`
 }
 
 // Permission defines model for Permission.
@@ -552,11 +554,11 @@ type PubKey struct {
 // ServiceAccount defines model for ServiceAccount.
 type ServiceAccount struct {
 	AccountId   string                      `json:"account_id"`
-	AppId       *string                     `json:"app_id"`
+	AppId       string                      `json:"app_id"`
 	Creds       *[]ServiceAccountCredential `json:"creds,omitempty"`
 	FirstParty  bool                        `json:"first_party"`
 	Name        string                      `json:"name"`
-	OrgId       *string                     `json:"org_id"`
+	OrgId       string                      `json:"org_id"`
 	Permissions []string                    `json:"permissions"`
 }
 
@@ -735,12 +737,12 @@ type ServicesReqCredentialUpdate struct {
 // ServicesReqServiceAccountsAccessToken defines model for _services_req_service-accounts_access-token.
 type ServicesReqServiceAccountsAccessToken struct {
 	AccountId string                                        `json:"account_id"`
-	AppId     *string                                       `json:"app_id"`
+	AppId     string                                        `json:"app_id"`
 	AuthType  ServicesReqServiceAccountsAccessTokenAuthType `json:"auth_type"`
 
 	// Service account token for auth_type="static_token"
 	Creds *ServicesServiceAccountsCredsStaticToken `json:"creds,omitempty"`
-	OrgId *string                                  `json:"org_id"`
+	OrgId string                                   `json:"org_id"`
 }
 
 // ServicesReqServiceAccountsAccessTokenAuthType defines model for ServicesReqServiceAccountsAccessToken.AuthType.
@@ -792,8 +794,8 @@ type ServicesResAuthorizeServiceTokenType string
 
 // ServicesResServiceAccountsAccessTokens defines model for _services_res_service-accounts_access-tokens.
 type ServicesResServiceAccountsAccessTokens struct {
-	AppId *string               `json:"app_id"`
-	OrgId *string               `json:"org_id"`
+	AppId string                `json:"app_id"`
+	OrgId string                `json:"org_id"`
 	Token SharedResRokwireToken `json:"token"`
 }
 
@@ -991,6 +993,7 @@ type SharedResAccount struct {
 	Preferences    *map[string]interface{}  `json:"preferences"`
 	Profile        *ProfileFields           `json:"profile,omitempty"`
 	Roles          *[]AppOrgRole            `json:"roles,omitempty"`
+	SystemConfigs  *map[string]interface{}  `json:"system_configs"`
 }
 
 // SharedResAccountCheck defines model for _shared_res_AccountCheck.
@@ -1124,11 +1127,11 @@ type SystemReqCreateAuthType_Params struct {
 // SystemReqCreateServiceAccount defines model for _system_req_create_service-account.
 type SystemReqCreateServiceAccount struct {
 	AccountId   *string                     `json:"account_id,omitempty"`
-	AppId       *string                     `json:"app_id"`
+	AppId       string                      `json:"app_id"`
 	Creds       *[]ServiceAccountCredential `json:"creds,omitempty"`
 	FirstParty  *bool                       `json:"first_party,omitempty"`
 	Name        *string                     `json:"name,omitempty"`
-	OrgId       *string                     `json:"org_id"`
+	OrgId       string                      `json:"org_id"`
 	Permissions *[]string                   `json:"permissions,omitempty"`
 }
 
@@ -1243,6 +1246,9 @@ type DeleteAdminApplicationAccountsIdRolesJSONBody AdminReqRevokeRolesFromAccoun
 
 // PutAdminApplicationAccountsIdRolesJSONBody defines parameters for PutAdminApplicationAccountsIdRoles.
 type PutAdminApplicationAccountsIdRolesJSONBody AdminReqGrantRolesToAccount
+
+// PutAdminApplicationAccountsIdSystemConfigsJSONBody defines parameters for PutAdminApplicationAccountsIdSystemConfigs.
+type PutAdminApplicationAccountsIdSystemConfigsJSONBody map[string]interface{}
 
 // PostAdminApplicationGroupsJSONBody defines parameters for PostAdminApplicationGroups.
 type PostAdminApplicationGroupsJSONBody AdminReqCreateApplicationGroup
@@ -1570,20 +1576,20 @@ type PostSystemServiceAccountsParams struct {
 type DeleteSystemServiceAccountsIdParams struct {
 
 	// Application ID of the service account instance
-	AppId *string `json:"app_id,omitempty"`
+	AppId string `json:"app_id"`
 
 	// Organization ID of the service account instance
-	OrgId *string `json:"org_id,omitempty"`
+	OrgId string `json:"org_id"`
 }
 
 // GetSystemServiceAccountsIdParams defines parameters for GetSystemServiceAccountsId.
 type GetSystemServiceAccountsIdParams struct {
 
 	// Application ID of the service account instance
-	AppId *string `json:"app_id,omitempty"`
+	AppId string `json:"app_id"`
 
 	// Organization ID of the service account instance
-	OrgId *string `json:"org_id,omitempty"`
+	OrgId string `json:"org_id"`
 }
 
 // PutSystemServiceAccountsIdJSONBody defines parameters for PutSystemServiceAccountsId.
@@ -1593,10 +1599,10 @@ type PutSystemServiceAccountsIdJSONBody SystemReqUpdateServiceAccount
 type PutSystemServiceAccountsIdParams struct {
 
 	// Application ID of the service account instance
-	AppId *string `json:"app_id,omitempty"`
+	AppId string `json:"app_id"`
 
 	// Organization ID of the service account instance
-	OrgId *string `json:"org_id,omitempty"`
+	OrgId string `json:"org_id"`
 }
 
 // DeleteSystemServiceAccountsIdCredsParams defines parameters for DeleteSystemServiceAccountsIdCreds.
@@ -1682,6 +1688,9 @@ type DeleteAdminApplicationAccountsIdRolesJSONRequestBody DeleteAdminApplication
 
 // PutAdminApplicationAccountsIdRolesJSONRequestBody defines body for PutAdminApplicationAccountsIdRoles for application/json ContentType.
 type PutAdminApplicationAccountsIdRolesJSONRequestBody PutAdminApplicationAccountsIdRolesJSONBody
+
+// PutAdminApplicationAccountsIdSystemConfigsJSONRequestBody defines body for PutAdminApplicationAccountsIdSystemConfigs for application/json ContentType.
+type PutAdminApplicationAccountsIdSystemConfigsJSONRequestBody PutAdminApplicationAccountsIdSystemConfigsJSONBody
 
 // PostAdminApplicationGroupsJSONRequestBody defines body for PostAdminApplicationGroups for application/json ContentType.
 type PostAdminApplicationGroupsJSONRequestBody PostAdminApplicationGroupsJSONBody